requirements.txt
asgiref==3.8.1
channels==3.0.4
channels-redis==4.2.0
daphne==3.0.2
Django==5.0.6
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
nltk==3.9.1
textblob==0.19.0
redis==5.2.0
sqlparse==0.5.0
psycopg2==2.9.9
pytz==2024.1
tzdata==2024.1
